Fix file length violations and address review findings
1. webhook_server.rs: bind_and_spawn now records actual bound address
   from listener.local_addr() into self.config.addr before spawning

2. capturing_store.rs: Split into mod.rs (111 lines) and delegation.rs
   (382 lines) to stay under 400-line limit. Added doc comments for
   record_status and record_event explaining job ID is discarded and
   only most recent call is retained.

3. null_database.rs: Split 620-line file into facade (44 lines) and
   per-store submodules:
   - job_store.rs (87 lines)
   - sandbox_store.rs (103 lines)
   - conversation_store.rs (133 lines)
   - routine_store.rs (92 lines)
   - tool_failure_store.rs (31 lines)
   - settings_store.rs (71 lines)
   - workspace_store.rs (118 lines)
