+++ src/testing/null_db/capturing_store/delegation.rs
@@ -1,12 +1,11 @@
-//! Capturing database wrapper for tests.
+//! Delegate implementations for CapturingStore.
 //!
-//! Provides a [`CapturingStore`] that wraps [`NullDatabase`] and captures
-//! specific method calls for test assertions.
-
-use std::sync::Arc;
+//! This module contains all the `delegate!` macro invocations that forward
+//! trait implementations to the inner NullDatabase. The CapturingStore
+//! overrides only `update_job_status` and `save_job_event` to capture calls;
+//! all other methods are delegated unchanged.
 
 use delegate::delegate;
-use tokio::sync::Mutex;
 use uuid::Uuid;
 
 use crate::agent::{Routine, routine::RoutineRun};
@@ -23,92 +22,7 @@ use crate::history::{
 };
 use crate::workspace::{MemoryChunk, MemoryDocument, SearchResult, WorkspaceEntry};
 
-use super::NullDatabase;
-
-/// Captured status update call.
-#[derive(Debug, Clone)]
-pub struct StatusCall {
-    /// The job status that was recorded.
-    pub status: JobState,
-    /// Optional failure reason associated with the status.
-    pub reason: Option<String>,
-}
-
-/// Captured job event call.
-#[derive(Debug, Clone)]
-pub struct EventCall {
-    /// The event type string (e.g., "result").
-    pub event_type: String,
-    /// The JSON data payload associated with the event.
-    pub data: serde_json::Value,
-}
-
-/// Thread-safe storage for captured calls.
-#[derive(Debug, Default)]
-pub struct Calls {
-    /// The last status update call captured, if any.
-    pub last_status: Mutex<Option<StatusCall>>,
-    /// The last event call captured, if any.
-    pub last_event: Mutex<Option<EventCall>>,
-}
-
-impl Calls {
-    /// Create a new empty Calls container.
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Record a status update call.
-    pub async fn record_status(&self, _id: Uuid, status: JobState, reason: Option<&str>) {
-        *self.last_status.lock().await = Some(StatusCall {
-            status,
-            reason: reason.map(ToOwned::to_owned),
-        });
-    }
-
-    /// Record an event call.
-    pub async fn record_event(
-        &self,
-        _job_id: Uuid,
-        event_type: SandboxEventType,
-        data: &serde_json::Value,
-    ) {
-        *self.last_event.lock().await = Some(EventCall {
-            event_type: event_type.as_str().to_string(),
-            data: data.clone(),
-        });
-    }
-}
-
-/// A database wrapper that captures calls to specific methods for testing.
-///
-/// Delegates all other methods to the inner [`NullDatabase`].
-#[derive(Debug)]
-pub struct CapturingStore {
-    inner: NullDatabase,
-    calls: Arc<Calls>,
-}
-
-impl CapturingStore {
-    /// Create a new capturing store with an inner NullDatabase.
-    pub fn new() -> Self {
-        Self {
-            inner: NullDatabase::new(),
-            calls: Arc::new(Calls::new()),
-        }
-    }
-
-    /// Access the captured calls for assertions.
-    pub fn calls(&self) -> &Arc<Calls> {
-        &self.calls
-    }
-}
-
-impl Default for CapturingStore {
-    fn default() -> Self {
-        Self::new()
-    }
-}
+use super::CapturingStore;
 
 impl crate::db::NativeDatabase for CapturingStore {
     delegate! {
@@ -226,7 +140,6 @@ impl crate::db::NativeSandboxStore for CapturingStore {
     }
 }
 
-// Delegate all other traits to inner NullDatabase
 impl crate::db::NativeConversationStore for CapturingStore {
     delegate! {
         to self.inner {
